feat: v0.0.1

+++ server/src/routes/storage.ts
@@ -13,6 +13,9 @@ import Robot from '../models/Robot';
 import Run from '../models/Run';
 import { BinaryOutputService } from '../storage/mino';
 import { workflowQueue } from '../worker';
+import { AuthenticatedRequest } from './record';
+import { computeNextRun } from '../utils/schedule';
+import { capture } from "../utils/analytics";
 
 export const router = Router();
 
@@ -57,11 +60,22 @@ router.get('/recordings/:id', requireSignIn, async (req, res) => {
 /**
  * DELETE endpoint for deleting a recording from the storage.
  */
-router.delete('/recordings/:id', requireSignIn, async (req, res) => {
+router.delete('/recordings/:id', requireSignIn, async (req: AuthenticatedRequest, res) => {
+  if (!req.user) {
+    return res.status(401).send({ error: 'Unauthorized' });
+  }
   try {
     await Robot.destroy({
       where: { 'recording_meta.id': req.params.id }
     });
+    capture(
+      'maxun-oss-robot-deleted',
+      {
+        robotId: req.params.id,
+        user_id: req.user?.id,
+        deleted_at: new Date().toISOString(),
+      }
+    )
     return res.send(true);
   } catch (e) {
     const { message } = e as Error;
@@ -86,9 +100,20 @@ router.get('/runs', requireSignIn, async (req, res) => {
 /**
  * DELETE endpoint for deleting a run from the storage.
  */
-router.delete('/runs/:id', requireSignIn, async (req, res) => {
+router.delete('/runs/:id', requireSignIn, async (req: AuthenticatedRequest, res) => {
+  if (!req.user) {
+    return res.status(401).send({ error: 'Unauthorized' });
+  }
   try {
     await Run.destroy({ where: { runId: req.params.id } });
+    capture(
+      'maxun-oss-run-deleted',
+      {
+        runId: req.params.id,
+        user_id: req.user?.id,
+        deleted_at: new Date().toISOString(),
+      }
+    )
     return res.send(true);
   } catch (e) {
     const { message } = e as Error;
@@ -101,7 +126,7 @@ router.delete('/runs/:id', requireSignIn, async (req, res) => {
  * PUT endpoint for starting a remote browser instance and saving run metadata to the storage.
  * Making it ready for interpretation and returning a runId.
  */
-router.put('/runs/:id', requireSignIn, async (req, res) => {
+router.put('/runs/:id', requireSignIn, async (req: AuthenticatedRequest, res) => {
   try {
     const recording = await Robot.findOne({
       where: {
@@ -112,6 +137,10 @@ router.put('/runs/:id', requireSignIn, async (req, res) => {
 
     if (!recording || !recording.recording_meta || !recording.recording_meta.id) {
       return res.status(404).send({ error: 'Recording not found' });
+    }
+
+    if (!req.user) {
+      return res.status(401).send({ error: 'Unauthorized' });
     }
 
     const proxyConfig = await getDecryptedProxyConfig(req.user.id);
@@ -127,18 +156,20 @@ router.put('/runs/:id', requireSignIn, async (req, res) => {
       };
     }
 
+    console.log(`Proxy config for run: ${JSON.stringify(proxyOptions)}`)
+
     const id = createRemoteBrowserForRun({
       browser: chromium,
       launchOptions: {
         headless: true,
         proxy: proxyOptions.server ? proxyOptions : undefined,
       }
-    });
+    }, req.user.id);
 
     const runId = uuid();
 
     const run = await Run.create({
-      status: 'RUNNING',
+      status: 'running',
       name: recording.recording_meta.name,
       robotId: recording.id,
       robotMetaId: recording.recording_meta.id,
@@ -186,9 +217,9 @@ router.get('/runs/run/:id', requireSignIn, async (req, res) => {
 /**
  * PUT endpoint for finishing a run and saving it to the storage.
  */
-router.post('/runs/run/:id', requireSignIn, async (req, res) => {
+router.post('/runs/run/:id', requireSignIn, async (req: AuthenticatedRequest, res) => {
   try {
-    console.log(`Params for POST /runs/run/:id`, req.params.id)
+    if (!req.user) { return res.status(401).send({ error: 'Unauthorized' }); }
 
     const run = await Run.findOne({ where: { runId: req.params.id } });
     if (!run) {
@@ -220,6 +251,40 @@ router.post('/runs/run/:id', requireSignIn, async (req, res) => {
         serializableOutput: interpretationInfo.serializableOutput,
         binaryOutput: uploadedBinaryOutput,
       });
+
+      let totalRowsExtracted = 0;
+      let extractedScreenshotsCount = 0;
+      let extractedItemsCount = 0;
+
+      if (run.dataValues.binaryOutput && run.dataValues.binaryOutput["item-0"]) {
+        extractedScreenshotsCount = 1;
+      }
+
+      if (run.dataValues.serializableOutput && run.dataValues.serializableOutput["item-0"]) {
+        const itemsArray = run.dataValues.serializableOutput["item-0"];
+        extractedItemsCount = itemsArray.length;
+
+        totalRowsExtracted = itemsArray.reduce((total, item) => {
+          return total + Object.keys(item).length;
+        }, 0);
+      }
+
+      console.log(`Extracted Items Count: ${extractedItemsCount}`);
+      console.log(`Extracted Screenshots Count: ${extractedScreenshotsCount}`);
+      console.log(`Total Rows Extracted: ${totalRowsExtracted}`);
+
+      capture(
+        'maxun-oss-run-created-manual',
+        {
+          runId: req.params.id,
+          user_id: req.user?.id,
+          created_at: new Date().toISOString(),
+          status: 'success',
+          totalRowsExtracted,
+          extractedItemsCount,
+          extractedScreenshotsCount,
+        }
+      )
       try {
         googleSheetUpdateTasks[plainRun.runId] = {
           robotId: plainRun.robotMetaId,
@@ -237,68 +302,56 @@ router.post('/runs/run/:id', requireSignIn, async (req, res) => {
     }
   } catch (e) {
     const { message } = e as Error;
+    // If error occurs, set run status to failed
+    const run = await Run.findOne({ where: { runId: req.params.id } });
+    if (run) {
+      await run.update({
+        status: 'failed',
+        finishedAt: new Date().toLocaleString(),
+      });
+    }
     logger.log('info', `Error while running a recording with id: ${req.params.id} - ${message}`);
+    capture(
+      'maxun-oss-run-created-manual',
+      {
+        runId: req.params.id,
+        user_id: req.user?.id,
+        created_at: new Date().toISOString(),
+        status: 'failed',
+        error_message: message,
+      }
+    );
     return res.send(false);
   }
 });
 
-router.put('/schedule/:id/', requireSignIn, async (req, res) => {
-  console.log(req.body);
+router.put('/schedule/:id/', requireSignIn, async (req: AuthenticatedRequest, res) => {
   try {
     const { id } = req.params;
-    const {
-      // enabled = true,
-      runEvery,
-      runEveryUnit,
-      startFrom,
-      atTimeStart,
-      atTimeEnd,
-      timezone
-    } = req.body;
+    const { runEvery, runEveryUnit, startFrom, dayOfMonth, atTimeStart, atTimeEnd, timezone } = req.body;
 
     const robot = await Robot.findOne({ where: { 'recording_meta.id': id } });
     if (!robot) {
       return res.status(404).json({ error: 'Robot not found' });
     }
 
-    // If disabled, remove scheduling
-    // if (!enabled) {
-    //   // Remove existing job from queue if it exists
-    //   const existingJobs = await workflowQueue.getJobs(['delayed', 'waiting']);
-    //   for (const job of existingJobs) {
-    //     if (job.data.id === id) {
-    //       await job.remove();
-    //     }
-    //   }
-
-    //   // Update robot to disable scheduling
-    //   await robot.update({
-    //     schedule: null
-    //   });
-
-    //   return res.status(200).json({
-    //     message: 'Schedule disabled successfully'
-    //   });
-    // }
-
-    if (!id || !runEvery || !runEveryUnit || !startFrom || !timezone || (runEveryUnit === 'HOURS' || runEveryUnit === 'MINUTES') && (!atTimeStart || !atTimeEnd)) {
+    // Validate required parameters
+    if (!runEvery || !runEveryUnit || !startFrom || !atTimeStart || !atTimeEnd || !timezone) {
       return res.status(400).json({ error: 'Missing required parameters' });
     }
 
-    if (!['HOURS', 'DAYS', 'WEEKS', 'MONTHS', 'MINUTES'].includes(runEveryUnit)) {
-      return res.status(400).json({ error: 'Invalid runEvery unit' });
-    }
-
+    // Validate time zone
     if (!moment.tz.zone(timezone)) {
       return res.status(400).json({ error: 'Invalid timezone' });
     }
 
+    // Validate and parse start and end times
     const [startHours, startMinutes] = atTimeStart.split(':').map(Number);
     const [endHours, endMinutes] = atTimeEnd.split(':').map(Number);
 
     if (isNaN(startHours) || isNaN(startMinutes) || isNaN(endHours) || isNaN(endMinutes) ||
-        startHours < 0 || startHours > 23 || startMinutes < 0 || startMinutes > 59 ||
-        endHours < 0 || endHours > 23 || endMinutes < 0 || endMinutes > 59) {
+      startHours < 0 || startHours > 23 || startMinutes < 0 || startMinutes > 59 ||
+      endHours < 0 || endHours > 23 || endMinutes < 0 || endMinutes > 59) {
       return res.status(400).json({ error: 'Invalid time format' });
     }
 
@@ -307,86 +360,94 @@ router.put('/schedule/:id/', requireSignIn, async (req, res) => {
       return res.status(400).json({ error: 'Invalid start day' });
     }
 
+    // Build cron expression based on run frequency and starting day
     let cronExpression;
+    const dayIndex = days.indexOf(startFrom);
+
     switch (runEveryUnit) {
       case 'MINUTES':
+        cronExpression = `${startMinutes} */${runEvery} * * *`;
+        break;
       case 'HOURS':
-        cronExpression = `${startMinutes}-${endMinutes} */${runEvery} * * *`;
+        cronExpression = `${startMinutes} */${runEvery} * * *`;
         break;
       case 'DAYS':
         cronExpression = `${startMinutes} ${startHours} */${runEvery} * *`;
         break;
       case 'WEEKS':
-        const dayIndex = days.indexOf(startFrom);
-        cronExpression = `${startMinutes} ${startHours} * * ${dayIndex}/${7 * runEvery}`;
+        cronExpression = `${startMinutes} ${startHours} * * ${dayIndex}`;
         break;
       case 'MONTHS':
-        cronExpression = `${startMinutes} ${startHours} 1-7 */${runEvery} *`;
+        // todo: handle leap year
+        cronExpression = `0 ${atTimeStart} ${dayOfMonth} * *`;
         if (startFrom !== 'SUNDAY') {
-          const dayIndex = days.indexOf(startFrom);
           cronExpression += ` ${dayIndex}`;
         }
         break;
+      default:
+        return res.status(400).json({ error: 'Invalid runEveryUnit' });
     }
 
+    // Validate cron expression
     if (!cronExpression || !cron.validate(cronExpression)) {
       return res.status(400).json({ error: 'Invalid cron expression generated' });
     }
 
-    const runId = uuid();
-    const userId = req.user.id;
+    if (!req.user) {
+      return res.status(401).json({ error: 'Unauthorized' });
+    }
 
-    // Remove existing jobs for this robot just in case some were left
-    // const existingJobs = await workflowQueue.getJobs(['delayed', 'waiting']);
-    // for (const job of existingJobs) {
-    //   if (job.data.id === id) {
-    //     await job.remove();
-    //   }
-    // }
-
-    // Add new job
+    // Create the job in the queue with the cron expression
     const job = await workflowQueue.add(
       'run workflow',
-      { id, runId, userId },
+      { id, runId: uuid(), userId: req.user.id },
       {
         repeat: {
           pattern: cronExpression,
-          tz: timezone
-        }
+          tz: timezone,
+        },
       }
     );
 
-    const nextRun = job.timestamp;
+    const nextRunAt = computeNextRun(cronExpression, timezone);
 
-    // Update robot with schedule details
     await robot.update({
       schedule: {
         runEvery,
         runEveryUnit,
         startFrom,
+        dayOfMonth,
         atTimeStart,
         atTimeEnd,
         timezone,
         cronExpression,
         lastRunAt: undefined,
-        nextRunAt: new Date(nextRun)
-      }
+        nextRunAt: nextRunAt || undefined,
+      },
     });
+
+    capture(
+      'maxun-oss-robot-scheduled',
+      {
+        robotId: id,
+        user_id: req.user.id,
+        scheduled_at: new Date().toISOString(),
+      }
+    )
 
     // Fetch updated schedule details after setting it
     const updatedRobot = await Robot.findOne({ where: { 'recording_meta.id': id } });
 
     res.status(200).json({
       message: 'success',
-      runId,
-      robot: updatedRobot
+      robot: updatedRobot,
     });
-
   } catch (error) {
     console.error('Error scheduling workflow:', error);
     res.status(500).json({ error: 'Failed to schedule workflow' });
   }
 });
+
 
 // Endpoint to get schedule details
 router.get('/schedule/:id', requireSignIn, async (req, res) => {
@@ -408,9 +469,13 @@ router.get('/schedule/:id', requireSignIn, async (req, res) => {
 });
 
 // Endpoint to delete schedule
-router.delete('/schedule/:id', requireSignIn, async (req, res) => {
+router.delete('/schedule/:id', requireSignIn, async (req: AuthenticatedRequest, res) => {
   try {
     const { id } = req.params;
+
+    if (!req.user) {
+      return res.status(401).json({ error: 'Unauthorized' });
+    }
 
     const robot = await Robot.findOne({ where: { 'recording_meta.id': id } });
     if (!robot) {
@@ -429,6 +494,15 @@ router.delete('/schedule/:id', requireSignIn, async (req, res) => {
     await robot.update({
       schedule: null
     });
+
+    capture(
+      'maxun-oss-robot-schedule-deleted',
+      {
+        robotId: id,
+        user_id: req.user?.id,
+        unscheduled_at: new Date().toISOString(),
+      }
+    )
 
     res.status(200).json({ message: 'Schedule deleted successfully' });
